requirements.txt
requests>=2.31.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pywin32>=306; platform_system=="Windows"
tkinter-tooltip>=2.1.0
